Refactor `HitlTypeSelection` into a standalone module for better modularity and reuse.
